+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/components/configuration_form/configuration_form.tsx
@@ -3,30 +3,57 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { EuiSpacer } from '@elastic/eui';
 
-import { Types } from '../../mappings_state';
-import { DynamicMappingForm } from './dynamic_mapping_form';
-import { SourceFieldForm } from './source_field_form';
+import { useForm, Form, SerializerFunc } from '../../shared_imports';
+import { Types, useDispatch } from '../../mappings_state';
+import { configurationFormSchema } from './configuration_form_schema';
+import { DynamicMapping } from './dynamic_mapping';
+import { SourceField } from './source_field';
 
 type MappingsConfiguration = Types['MappingsConfiguration'];
-type SourceField = Types['SourceField'];
 
 interface Props {
-  configurationDefaultValue?: MappingsConfiguration;
-  sourceFieldDefaultValue?: SourceField;
+  defaultValue?: MappingsConfiguration;
 }
 
-export const ConfigurationForm = React.memo(
-  ({ configurationDefaultValue, sourceFieldDefaultValue }: Props) => {
-    return (
-      <>
-        <DynamicMappingForm defaultValue={configurationDefaultValue} />
-        <EuiSpacer size="xl" />
-        <SourceFieldForm defaultValue={sourceFieldDefaultValue} />
-      </>
-    );
-  }
-);
+const formSerializer: SerializerFunc<MappingsConfiguration> = formData => {
+  const {
+    enabled: dynamicMappingsEnabled,
+    throwErrorsForUnmappedFields,
+    ...configurationData
+  } = formData;
+
+  const dynamic = dynamicMappingsEnabled ? true : throwErrorsForUnmappedFields ? 'strict' : false;
+
+  return {
+    ...configurationData,
+    dynamic,
+  };
+};
+
+export const ConfigurationForm = React.memo(({ defaultValue }: Props) => {
+  const { form } = useForm<MappingsConfiguration>({
+    schema: configurationFormSchema,
+    serializer: formSerializer,
+    defaultValue,
+  });
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const subscription = form.subscribe(updatedConfiguration => {
+      dispatch({ type: 'configuration.update', value: updatedConfiguration });
+    });
+    return subscription.unsubscribe;
+  }, [form]);
+
+  return (
+    <Form form={form}>
+      <DynamicMapping />
+      <EuiSpacer size="xl" />
+      <SourceField />
+    </Form>
+  );
+});
