--- x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/components/configuration_form/source_field_form/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export { SourceFieldForm } from './source_field_form';
