+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/mappings_editor.tsx
@@ -30,13 +30,12 @@ export const MappingsEditor = React.memo(
   ({ onUpdate, defaultValue = {}, indexSettings }: Props) => {
     const [selectedTab, selectTab] = useState<'fields' | 'advanced'>('fields');
 
-    const { configurationDefaultValue, sourceFieldDefaultValue, fieldsDefaultValue } = useMemo(
+    const { configurationDefaultValue, fieldsDefaultValue } = useMemo(
       () => ({
-        configurationDefaultValue: pick(
-          defaultValue,
-          CONFIGURATION_FIELDS
-        ) as Types['MappingsConfiguration'],
-        sourceFieldDefaultValue: defaultValue._source || {},
+        configurationDefaultValue: {
+          ...(pick(defaultValue, CONFIGURATION_FIELDS) as Types['MappingsConfiguration']),
+          _source: defaultValue._source || {},
+        },
         fieldsDefaultValue: defaultValue.properties || {},
       }),
       [defaultValue]
@@ -61,10 +60,7 @@ export const MappingsEditor = React.memo(
                   {editor}
                 </>
               ) : (
-                <ConfigurationForm
-                  configurationDefaultValue={configurationDefaultValue}
-                  sourceFieldDefaultValue={sourceFieldDefaultValue}
-                />
+                <ConfigurationForm defaultValue={configurationDefaultValue} />
               );
 
             return (
