+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/mappings_state.tsx
@@ -12,22 +12,19 @@ import {
   MappingsConfiguration,
   MappingsFields,
   State,
-  SourceField,
   Dispatch,
 } from './reducer';
 import { Field, FieldsEditor } from './types';
 import { normalize, deNormalize, canUseMappingsEditor } from './lib';
 
-type Mappings = MappingsConfiguration &
-  SourceField & {
-    properties: MappingsFields;
-  };
+type Mappings = MappingsConfiguration & {
+  properties: MappingsFields;
+};
 
 export interface Types {
   Mappings: Mappings;
   MappingsConfiguration: MappingsConfiguration;
   MappingsFields: MappingsFields;
-  SourceField: SourceField;
 }
 
 export interface OnUpdateHandlerArg {
@@ -60,13 +57,6 @@ export const MappingsState = React.memo(({ children, onUpdate, defaultValue }: P
   const initialState: State = {
     isValid: undefined,
     configuration: {
-      data: {
-        raw: {},
-        format: () => ({} as Mappings),
-      },
-      validate: () => Promise.resolve(true),
-    },
-    sourceField: {
       data: {
         raw: {},
         format: () => ({} as Mappings),
@@ -122,18 +112,6 @@ export const MappingsState = React.memo(({ children, onUpdate, defaultValue }: P
           }
         }
 
-        const {
-          enabled: dynamicMappingsEnabled,
-          throwErrorsForUnmappedFields,
-          ...configurationData
-        } = nextState.configuration.data.format();
-
-        const dynamicMapping = dynamicMappingsEnabled
-          ? true
-          : throwErrorsForUnmappedFields
-          ? 'strict'
-          : false;
-
         // Pull the mappings properties from the current editor
         const fields =
           nextState.documentFields.editor === 'json'
@@ -141,11 +119,7 @@ export const MappingsState = React.memo(({ children, onUpdate, defaultValue }: P
             : deNormalize(nextState.fields);
 
         return {
-          dynamic: dynamicMapping,
-          ...configurationData,
-          _source: {
-            ...nextState.sourceField.data.format(),
-          },
+          ...nextState.configuration.data.format(),
           properties: fields,
         };
       },
