+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/reducer.ts
@@ -23,12 +23,11 @@ export interface MappingsConfiguration {
   date_detection: boolean;
   numeric_detection: boolean;
   dynamic_date_formats: string[];
-}
-
-export interface SourceField {
-  enabled?: boolean;
-  includes?: string[];
-  excludes?: string[];
+  _source: {
+    enabled?: boolean;
+    includes?: string[];
+    excludes?: string[];
+  };
 }
 
 export interface MappingsFields {
@@ -47,7 +46,6 @@ interface DocumentFieldsState {
 export interface State {
   isValid: boolean | undefined;
   configuration: OnFormUpdateArg<MappingsConfiguration>;
-  sourceField: OnFormUpdateArg<SourceField>;
   documentFields: DocumentFieldsState;
   fields: NormalizedFields;
   fieldForm?: OnFormUpdateArg<any>;
@@ -59,7 +57,6 @@ export interface State {
 
 export type Action =
   | { type: 'configuration.update'; value: OnFormUpdateArg<MappingsConfiguration> }
-  | { type: 'sourceField.update'; value: OnFormUpdateArg<SourceField> }
   | { type: 'fieldForm.update'; value: OnFormUpdateArg<any> }
   | { type: 'field.add'; value: Field }
   | { type: 'field.remove'; value: string }
@@ -229,17 +226,6 @@ export const reducer = (state: State, action: Action): State => {
       const nextState = {
         ...state,
         configuration: action.value,
-      };
-
-      const isValid = isStateValid(nextState);
-      nextState.isValid = isValid;
-
-      return nextState;
-    }
-    case 'sourceField.update': {
-      const nextState = {
-        ...state,
-        sourceField: action.value,
       };
 
       const isValid = isStateValid(nextState);
