+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/shared_imports.ts
@@ -35,6 +35,7 @@ export {
   TextAreaField,
   TextField,
   ToggleField,
+  ComboBoxField,
 } from '../../../../../../../../src/plugins/es_ui_shared/static/forms/components';
 
 export {
